Refactors tree and navigator data structure.
